LayoutTests/platform/ios/css3/flexbox/flexitem-expected.txt


PASS .flexbox 1
PASS .flexbox 2
PASS .flexbox 3
PASS .flexbox 4
PASS .flexbox 5
PASS .flexbox 6
PASS .flexbox 7
PASS .flexbox 8
PASS .flexbox 9
PASS .flexbox 10
PASS .flexbox 11
PASS .flexbox 12
PASS .flexbox 13
PASS .flexbox 14
PASS .flexbox 15
PASS .flexbox 16
FAIL .flexbox 17 assert_equals:
<div class="flexbox">
  <input data-expected-display="block" data-expected-width="75" type="checkbox" value="radio">
  <input data-expected-display="block" data-expected-width="75" type="file" value="file">
  <input data-expected-display="block" data-expected-width="75" type="image" value="image">
  <input data-expected-display="block" data-expected-width="75" type="password" value="password">
  <input data-expected-display="block" data-expected-width="75" type="radio" value="radio">
  <input data-expected-display="block" data-expected-width="75" type="reset" value="reset">
  <input data-expected-display="block" data-expected-width="75" type="submit" value="submit">
  <input data-expected-display="block" data-expected-width="75" type="text" value="text">
</div>
width expected 75 but got 69
FAIL .flexbox 18 assert_equals:
<div class="flexbox" style="height:200px">
  <button data-expected-display="block" data-expected-width="100" data-expected-height="200">button</button>
  <canvas data-expected-display="block" data-expected-width="100" data-expected-height="200">canvas</canvas>
  <iframe data-expected-display="block" data-expected-width="100" data-expected-height="200" src="data:text/html,&lt;body bgcolor=#fff&gt;iframe&lt;/body&gt;"></iframe>
  <object data-expected-display="block" data-expected-width="100" data-expected-height="200">object</object>
  <select data-expected-display="block" data-expected-width="100" data-expected-height="200">
    <option>select</option>
  </select>
  <textarea data-expected-display="block" data-expected-width="100" data-expected-height="200">textarea</textarea>
</div>
width expected 100 but got 98
button

object




button



button

object





Some text
Some more text
